Moved code to a separate file to abstract some functionality.
